fix evaluateForce export for experiment
